8-BIT LOAD INSTRUCTIONS

+++ src/GameBoy/include/CPU/GB_Bus.h
@@ -3,13 +3,13 @@
 
 typedef enum {Start, Stop, Reset, Quit} ShellAction;
 
-uint8_t GB_BusRead(uint16_t * memory, uint16_t address)
+uint8_t GB_BusRead(uint8_t * memory, uint16_t address)
 {
     //based on address value fetch the mapped memory
     return 0xFF;
 }
 
-void GB_BusWrite(uint16_t * memory,uint16_t address,uint8_t value)
+void GB_BusWrite(uint8_t * memory, uint16_t address, uint8_t value)
 {
 
 }
